pyproject.toml
[tool.poetry]
name = "test-hq"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
flake8 = "^7.0.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.3.0"
pre-commit = "^3.6.2"
autoflake = "^2.3.0"
pyupdate = "^1.4.0"
reorder-python-imports = "^3.12.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
